Started restructuring the player module.
I've decided to split the player instance into three separate ones:
PlayerJoining, PlayerInRoom, and PlayerClose. The reason for this
is so that instances that use them know exactly how they will behave
instead of assuming the stage of connection a player is at.

I've completed the PlayerClose instance, but have yet to write the
other two, so currently the program will not compile.