Added the PlayerJoining instance.
Does the same thing as before, where it waits for the player's initial
request, but this time the code is a lot nicer! \o/

It now also sends the request to the lobby, and I can confirm the
send_close task (prev. the PlayerClose instance) works!

Although, I will need to rethink what level certain logs are printed
at, as the trace level has a TON of debugging info now.