Changed room notification channel from mpsc to broadcast.
This was due to a potential issue that I noticed, where the player
task could be waiting to send a request to the room via the mpsc
channel (if the channel is full), and that the room could be waiting
to send a notification to the player at the same time, causing a
gridlock of sorts.

Using a broadcast channel, you do not need to .await when sending
a value through the channel, which solves the gridlock problem.

However, there is now another error that could occur - the receiver
for notifications may lag behind and miss a value if it has not read
the channel in a while. If this happens, the connection is closed
with a generic error close code.