Now rely on reverse proxy to limit requests and connections.
Standing on the shoulders of giants, if it ain't broke don't fix it,
etc. etc.

When using the server in a production context, it's also likely that
there will be a reverse proxy in place, so this not only uses their
battle-tested implementation, but it also means I don't have to
code it manually! \o/

I'll be mentioning this in the README so users are at least aware.