Now use panic!() in specific situations.
Specifically, when breaking out of the loop would result in a hang,
since the task would be waiting on another task that is functioning
normally.

This should be okay, as it should be impossible for certain channels
to be dropped, as they are always in scope where these panic!()
calls are made.